Move sql files and utils to folders
